renamed packe windows to window Moved sidebar code to new Sidebar class
