+++ jsettlers.mapcreator/src/jsettlers/mapcreator/main/window/icons/icon-name-conventions.txt
