+++ jsettlers.mapcreator/src/jsettlers/mapcreator/main/window/icons/map-settings.png
