+++ jsettlers.mapcreator/src/jsettlers/mapcreator/main/window/icons/new.png
