+++ jsettlers.mapcreator/src/jsettlers/mapcreator/main/window/icons/redo.png
