+++ jsettlers.mapcreator/src/jsettlers/mapcreator/main/window/icons/save.png
