+++ jsettlers.mapcreator/src/jsettlers/mapcreator/main/window/icons/statistic.png
