+++ jsettlers.mapcreator/src/jsettlers/mapcreator/main/window/icons/tango-icon.txt
