+++ jsettlers.mapcreator/src/jsettlers/mapcreator/main/window/icons/undo.png
