+++ jsettlers.mapcreator/src/jsettlers/mapcreator/main/window/shortcut.properties
